Migrate money fields to decimal with per-currency rounding digits (ADR-0014 amended)
StockItem.Price/Factor, SaleLine.UnitPrice, Invoice.Total/Tax move from
double/long to decimal. InvoiceCalculator rounds to a new
CompanySettings.CurrencyDecimalDigits setting instead of a hardcoded whole
unit, defaulting to 0 to match the owner's main market (Paraguay, PYG,
no minor units).

+++ StockManagement.Api/Features/Invoices/InvoiceResponse.cs
@@ -5,13 +5,13 @@ namespace StockManagement.Api.Features.Invoices;
 
 
 /// <summary>
-/// Stored invoice; amounts in whole currency units
+/// Stored invoice; amounts rounded to the company's configured currency digits
 /// </summary>
-public sealed record InvoiceResponse(int Number, DateTime Date, DateTime ExpirationDate, long Total, long Tax, SaleCondition SaleCondition, int CustomerId, string CustomerName, IReadOnlyList<InvoiceLineResponse> Lines)
+public sealed record InvoiceResponse(int Number, DateTime Date, DateTime ExpirationDate, decimal Total, decimal Tax, SaleCondition SaleCondition, int CustomerId, string CustomerName, IReadOnlyList<InvoiceLineResponse> Lines)
 {
 	public static InvoiceResponse From(Invoice invoice)
 	{
-		var lines = (invoice.Items ?? []).Select(item => new InvoiceLineResponse(item.StockItem.Code, item.StockItem.Name, item.Amount, (decimal)item.StockItem.Price)).ToList();
+		var lines = (invoice.Items ?? []).Select(item => new InvoiceLineResponse(item.StockItem.Code, item.StockItem.Name, item.Amount, item.StockItem.Price)).ToList();
 		var customerName = invoice.Customer is Customer customer ? string.Join(" ", customer.Name, customer.Lastname).Trim() : "";
 		return new(invoice.Number, invoice.Date, invoice.ExpirationDate, invoice.Total, invoice.Tax, invoice.SaleCondition, invoice.Customer?.CustomerId ?? 0, customerName, lines);
 	}
